error handling: create/delete user

+++ src/repository/user-repository.js
@@ -16,6 +16,12 @@ class userRepository {
   }
 
   async createUser(User) {
+    const existingUser = await this.getUserByName(User.nameUser);
+    if (existingUser) {
+      console.log(existingUser);
+      throw new Error('User already exists');
+    }
+
     const sql = `INSERT INTO users (nameUser, passwordUser) VALUES (?, ?)`;
     const values = [User.nameUser, User.passwordUser];
     return new Promise((resolve, reject) => {
@@ -32,6 +38,7 @@ class userRepository {
   async updateUser(id, user) {
     const sql = `UPDATE users SET nameUser = ?, passwordUser = ? WHERE id = ?`;
     const values = [user.nameUser, user.passwordUser, id];
+
     return new Promise((resolve, reject) => {
       connection.query(sql, values, (err, result) => {
         if (err) {
@@ -44,8 +51,29 @@ class userRepository {
   }
 
   async deleteUser(id) {
-    const sql = `DELETE FROM users WHERE id = ?`;
-    const values = [id];
+    const existingUser = await this.getUserById(id);
+    if (existingUser && existingUser.length > 0) {
+      const sql = `DELETE FROM users WHERE id = ?`;
+      const values = [id];
+
+      return new Promise((resolve, reject) => {
+        connection.query(sql, values, (err, result) => {
+          if (err) {
+            reject(err);
+          } else {
+            resolve(result);
+          }
+        });
+      });
+    } else {
+      throw new Error('User not found');
+    }
+  }
+
+  async getUserByName(nameUser) {
+    const sql = `SELECT * FROM users WHERE nameUser = ?`;
+    const values = [nameUser];
+
     return new Promise((resolve, reject) => {
       connection.query(sql, values, (err, result) => {
         if (err) {
@@ -57,9 +85,12 @@ class userRepository {
     });
   }
 
-  async getUserByName(nameUser) {
-    const sql = `SELECT * FROM users WHERE nameUser = ?`;
-    const values = [nameUser];
+  async getUserById(id) {
+    const sql = `SELECT * FROM users WHERE id = ?`;
+    const values = [id];
+    if (!id) {
+      throw new Error('ID is required');
+    }
     return new Promise((resolve, reject) => {
       connection.query(sql, values, (err, result) => {
         if (err) {
